requirements.txt
certifi==2020.12.5
chardet==4.0.0
clarifai==2.6.2
click==7.1.2
configparser==3.8.1
Flask==1.1.2
flask-googlemaps==0.4.0
future==0.18.2
googleapis-common-protos==1.53.0
grpcio==1.37.0
htmlmin==0.1.12
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsmin==2.2.2
jsonschema==2.6.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.4
MarkupSafe==1.1.1
mkdocs==1.1.2
mkdocs-material==5.5.14
mkdocs-material-extensions==1.0.1
mkdocs-minify-plugin==0.3.0
nltk==3.6.2
protobuf==3.15.8
Pygments==2.9.0
PyJWT==1.7.1
pymdown-extensions==8.1.1
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
tornado==6.1
tqdm==4.60.0
twilio==6.57.0
urllib3==1.26.4
Werkzeug==1.0.1
